move getClientWithJson
