+++ src/sheettools.go
@@ -2,9 +2,12 @@ package main
 
 import (
 	"fmt"
+	"io/ioutil"
 	"net/http"
+	"os"
 	"time"
 
+	"golang.org/x/oauth2/google" // to get sheet client
 	"google.golang.org/api/sheets/v4"
 	"google.golang.org/appengine" // Required external App Engine library
 	"google.golang.org/appengine/log"
@@ -20,6 +23,25 @@ func getSheetClient(r *http.Request) (*sheets.Service, error) {
 		return nil, fmt.Errorf("unable to retrieve Sheets Client %v", err)
 	}
 	return srv, nil
+}
+
+func getClientWithJson(r *http.Request) *http.Client {
+	// リクエストからcontextを作成
+	// GAE log
+	ctx := appengine.NewContext(r)
+
+	credentialFilePath := "myfinance-01-dc1116b8f354.json"
+	data, err := ioutil.ReadFile(credentialFilePath)
+	if err != nil {
+		log.Errorf(ctx, "Unable to read client secret file: %v", err)
+		os.Exit(0)
+	}
+	conf, err := google.JWTConfigFromJSON(data, "https://www.googleapis.com/auth/spreadsheets")
+	if err != nil {
+		log.Errorf(ctx, "Unable to parse client secret file to config: %v", err)
+		os.Exit(0)
+	}
+	return conf.Client(ctx)
 }
 
 func getSheetData(r *http.Request, srv *sheets.Service, sheetId string, readRange string) [][]interface{} {
